+++ src/api/Ambs.Reporting/Ambs.Reporting.Service/Interfaces/IMetaDataService.cs
@@ -4,6 +4,6 @@ namespace Ambs.Reporting.Service.Interfaces;
 public interface IMetaDataService
 {
     IEnumerable<MetaDatum> GetAll();
-    MetaDatum Get(int id);
+    MetaDatum Get(long id);
     MetaDatum Save(MetaDatum dashboard);
 }
